backend/requirements.txt
asgiref==3.7.2
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==2.0.12
coreapi==2.3.3
coreschema==0.0.4
cryptography==41.0.4
defusedxml==0.7.1
Django==3.2.16
django-templated-mail==1.1.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.2
djoser==2.1.0
idna==3.4
importlib-metadata==1.7.0
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.3
oauthlib==3.2.2
Pillow==9.3.0
pycparser==2.21
PyJWT==2.1.0
python3-openid==3.2.0
pytz==2023.3.post1
requests==2.26.0
requests-oauthlib==1.3.1
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.4.2
sqlparse==0.4.4
typing_extensions==4.7.1
uritemplate==4.1.1
urllib3==1.26.16
zipp==3.15.0
gunicorn==20.1.0
psycopg2-binary==2.9.5
django-debug-toolbar==3.2.4
python-dotenv==0.21.1
django-filter==23.3
